+++ NEWS.md
@@ -1,8 +1,16 @@
+# correspondenceTables 0.7.4
+
+- aligning to the second round of comments from CRAN 
+- replacing `readr::write_excel_csv()` with `data.table::fwrite()`
+- removing absolute paths from vignettes and examples
+- the example of `newCorrespondenceTable()` using only the 100 hundred rows of the classifications
+- more details in the error messages 
+
 # correspondenceTables 0.7.3
 
 - aligning to the comments from CRAN 
-- using 'txtProgressBar()' and 'setTxtProgressBar()' functions
-- replacing 'cat()' with 'message()' function
+- using `txtProgressBar()` and `setTxtProgressBar()` functions
+- replacing `cat()` with `message()` function
 - examples updated to be able to test it automatically 
 
 # correspondenceTables 0.7.2